+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/cervel
 
+## 0.1.30
+
+### Patch Changes
+
+- @vercel/backends@0.8.22
+
 ## 0.1.29
 
 ### Patch Changes